ISPN-960 - Non serializable object errors not obvious when repl queue on
* Added an interceptor to check for marshalling errors early on when the
marshalling happens in a separate thread, i.e. repl queue, async marshalling
...etc.
* Checking whether an object is marshallable has been revamped since previous
method could not detect Serializable objects that might have refs to
non-serializable objects. Marshallable check now involves trying to marshall
an object of a type and the result is stored in a weakhashmap where the
result of whether a type is marshallable or not is stored.
* Finally, tests that relied on serialization counts have been adjusted
accordingly.