ISPN-1037 Test marshalling of different versions of classes
It uses a combination of a selected class name classloader and
Javassist in order to load new versions of classes defined in the test.
This enables creating payloads with original classes and then seeing
if new versions can unmarshall them correctly as well as new instances
of these new class versions.