ISPN-1239 - ConcurrentOverlappingLeaveTest failing with the graceful shutdown workaround
Let's say the initial cluster members are {A, B, C, D}, numOwners = 3, and D and C leave in quick succession.
With initial owners(k) = {B, C, D}, when D leaves B expects C to push the key to A.
If but C dies before pushing it, on the following rehash B doesn't push the key to A.

The solution is to remember the last CH for which rehashing completed successfully and base every rehash on the last successful CH.

Another problem is that pending transactions should not be allowed if a rehash is interrupted until the next rehash has completed.