ISPN-1106 - Rehashing into a running cluster causes shared processing lock contention
I added timeouts when acquiring the tx lock for read before executing commands in TransactionLoggerImpl.
For remote commands the timeout is set to 0, as the caller already holds the tx lock on the origin node and the rehash process already has the tx lock (for write) on the remote node, causing a deadlock.
I also added a timeout when sending the NODE_PUSH_COMPLETED command.