ISPN-1404 ISPN-1408 ISPN-1476 Combined commit for these jiras
* The following description provides an overview of this commit
that should help reviewers.
* ISPN-1404: This is about making hash seed configurable so that
Hot Rod clients can hash on the information available to them.
Changes in core/ module are related to this.
* ISPN-1408: The rest of code changes are pretty much related to
the introduction of Hot Rod protocol version 1.1 in order to make
Hot Rod topology information maintenance scalable.
* This includes transforming Hot Rod server to have versioned
encoders and get rid of all the old topology cache handling code.
* This also includes a big reshuffle of code in the Hot Rod client
which its current design was not really geared towards supporting
multiple versions of the Hot Rod protocol.
* ISPN-1476: Fix a very important bug in Hot Rod client's
ConsistentHashV1 which lead to clients not hashing the same way
as servers. As a result of this, CSAIntegrationTest was re-enabled.