ISPN-1483 - Don't clear the locks when the prepare fails with optimistic locking
This change is only to make optimistic locking and pessimistic the same, the behaviour didn't change.
With pessimitic locking we can't clear the locks in the prepare command because that will prevent the rollback command from running on the remote nodes.
With optimistic locking it doesn't matter because the transaction will have at least one modification in it, forcing the rollback command to run on the remote nodes.