ISPN-1529 Fix autoCommit configuration
* First, Remove the autoCommit getter and change variable to Boolean.
The schema generator was getting confused and assuming that the XML
attribute should be derived from the getter. That was cos the instance
variable was using boolean instead of Boolean.
* Secondly, autoCommit will only be detected if marked as overriden.
Apart from fixing this by creating a setter, transform XML file parsing
test in order to test overrides for named cache configuration tests.