ISPN-1606 - Data inconsistencies when a cache view installation is rolled back
Block new transactions the moment we see a leaver or on cache view prepare, unblock only on cache view commit.
Throw an exception if we get a cache view prepare after we stopped the cache.

We also need to fail a prepare if the response is null. It happened once in the test suite, and there was no exception to check either.