ISPN-1872 - Coordinator hangs when cache is loaded to it and l1cache enabled in cluster
The demo used a single single-threaded executor to execute both the user
actions and the table updates for cache entry notifications. Because
the single thread was busy with the PutMapCommand, the task queue grew
extremely large and in turn made the PutMapCommand take a lot longer.

Even after optimizing the executor the PutMapCommand still takes a very
long time to execute on remote nodes (>30 seconds), so I split the
generated map in pieces of in partitions of 1000 entries max.