HHH-7256 Use minimalPuts in Infinispan
putFromLoad now checks for minimalPuts to verify whether the entry
is already present in the cache. By doing this, Hibernate stats
appear more precise since putForExternalReads that are no-op because
the cache already contains the entry are not counted as puts. This
avoids confusion amongst users. This should be reverted to checking
PFER's return value when https://issues.jboss.org/browse/ISPN-1986
has been implemented.