Create a new SpringAsynchronousCache that will make use of asynchronous methods when the CacheManager is configured with useAsynchronousCacheOperations=true.
Using Asynchronous operations is desirable for many cases where Infinispan is used to memoize a method call with Spring's @Cacheable annotation, but the current implementation used only the synchronous methods. One side effect is that method calls for a cache put would block when trying to acquire a lock - even if the cache is configured to behave asynchronously - and any exceptions thrown in cache operations would propogate to the method that called the @Cacheable method.