ISPN-1174 Fix async cache store consistency issues with some operations
* This fix addresses issues related to interaction between remove,
eviction (with and without passivation) and async cache store, making
sure that the cache returns consistent information.
* To build unit tests I've had to patch internal cache components to
behave in a particular way, which has lead me to converting the
component metadata repository to a non-static class. This helps
enormously in building such unit tests without having to change
existing components.