ISPN-2372 DummyInMemoryCacheStore shouldn't hold MarshalledValues
Because the MarshalledValue keeps a reference to its marshaller,
it can't be used after the cache has been stopped.

I changed DummyInMemoryCacheStore to serialize its entries to a byte[].
The store also keeps a reference to the marshaller, so I changed
CacheLoaderFunctionalTest and PassivatePersistentTest to use the new store
after a cache restart.