ISPN-2301 Grouped keys across caches do not have the same ownership information
I created an alternative consistent hash factory that uses the JGroups
addresses to keep the CHs of different caches in sync. It's not the default,
so it needs to be configured explicitly with
<hash factory="org.infinispan.distribution.ch.SyncConsistentHashFactory"/>

In order to ignore the order of nodes in the members list, I made Address
implement Comparable - the JGroups Address was already Comparable.