ISPN-2398 StaleTransactionCleanupService doesn't unlock keys correctly for optimistic caches
Now that we send both the Commit/Rollback and the TxCompletionNotification
to the nodes where the tx has prepared, there's no need to this service to
release locks for the keys that no longer map to this node. These locks will
be released when the Commit/Rollback is received by this node (or if the
originator fails, the StaleTransactionCleanupService will do it).