ISPN-2373 Need to avoid deadlock between tasks
It is possible that a task from async executor starts another task also on async executor pool then wait for termination of this second task. This leads to deadlock if the pool has no free threads and the second task gets queued. This situation appears in ClusterTopologyManagerImpl and can be solved by not starting a second parallel task if not really necessary. Such cases are triggerRebalance() and execution of the command on local node in executeOnClusterSync().