ISPN-2300 Revert changed flag for entries where conditional ops failed
* Changed cache entry flagged was overloaded to mean that both that a
new entry had to be stored in the cache, and that a entry loaded from
the cache store had to be stored in the cache.
* Changed now represents a change, so a cache entry whose value in
memory is going to change as a direct result of the cache operation,
i.e. when the conditional operation succeeds.
* Loaded means that a entry was loaded from the cache store and needs
to be stored in the cache, i.e. when a conditional operation requires
comparing the old value which is present in the cache store, and if
the condition fails, the value in the cache store still needs updating
in the cache.