ISPN-2825 ClusterTopologyManagerImpl should not hold a lock while invoking an RPC
If the REBALANCE_START command takes a long time to be processed on
one of the nodes, the REBALANCE_CONFIRM commands from the rest of
the nodes are blocked and end up filling the OOB thread pool.

Add two new tests for large clusters.
Move CacheCreationStressTest to the profiling group.