ISPN-2875 Intermittent test failure: org.infinispan.statetransfer.DistNonTxOperationsDuringStateTransferTest
Delaying execution of Transport.invokeRemotely(CacheTopologyControlCommand) on coordinator is usually enough
to delay the start of rebalance on the joiner in most cases but not in the rare case when the joiner receives a non-null
pending CH right in the initial topology returned to it from the join call. This was causing the assert to fail. A more manageable
way of delaying start of rebalance is to override the RebalancePolicy rather than the Transport. This approach ensures (at least in this test scenario) that the initial topology has a null pending CH.