ISPN-3033 Lucene index stored in a StringBased CacheStore can get an unaligned bufferSize
There is no test as there was a test already; just that the existing test wouldn't
catch this case because the bufferSize wasn't part of the equals contract: the
assertion used the key.equals to verify correct interpretation.
Fixing the equals activated the unit test: org.infinispan.lucene.Key2StringMapperTest

NoopSegmentReadLockerTest had to be fixed too as there was a mistake which wasn't spotted
as it also took advantage of the same equals flaw.