ISPN-3217 Rebalance doesn't store data into cache store
In non-tx mode, if the originator of a put command is an owner, it will
only write the entry when the command is forwarded back to it by the
primary owner. But this rule shouldn't apply to state transfer puts,
which use the CACHE_MODE_LOCAL and SKIP_OWNERSHIP_CHECK flags.

DistCacheStoreInterceptor is now used for replicated caches as well.