ISPN-3172 Define and manage indexes and schemas via JMX
Refactored all protobuf serialization context stuff, properly injected it into the field bridge, added a MBean for it so new proto files can be registered. The protobuf metadata is kept in a replicated cache.

Also fixes ISPN-3523, by lifting the marshalling and query construction limitations that made it imposible for remote query to interoperate with embedded query.