ISPN-3670 Any commands received before the initial topology is installed should block
Because the initial cache topology is installed in a single phase,
it's possible to receive a StateRequestCommand from another node before
receiving the JOIN response from the coordinator and installing the initial
cache topology. As such, commands received before we have a cache topology
should block instead of being ignored.