Fix for race condition on clustered cache start
If an InvocationContext is created for a command before this factory instance has been started, it gets a null Equivalence for the key which later causes a NullPointerException in EquivalentHashMap.

This causes a key to be left locked, causing timeouts later.

In this commit we throw an IllegalStateException to prevent the caller from attempting the command, thus not acquiring the lock, thus avoiding the issue.

https://issues.jboss.org/browse/ISPN-3873