ISPN-4860 Separate previous value reading for Hot Rod 1.x and 2.x
* Each major protocol version has different expectations on
whether the previous value is present or not. The key difference,
as indicated in the documentation, is that when no previous value
is to be returned, not even the length is provided in 2.x. Before,
in 1.x, the length was there regardless. 2.x enhances processing
but using fully dedicated status codes for when the previous value
is returned.
* Added further tests for Hot Rod 1.x protocol to make sure
code remains backwards compatible.