ISPN-5150 In testing, near cache events should fire after event
* Before this fix, events were being fired before the near cache had
been updated, and as a result of that, client code was making working
on assumptions that might not have yet been applied in a multi threaded
environment.
* Instead of using the Hot Rod async thread pool for client listeners,
created a simple cached pool where each thread is associated with
a client listener, and the thread name is adjusted when the client
listener starts.
* Added multiple TRACE log messages to help debugging.