ISPN-5204 Don't include old schema files in the documentation and jars/distribution
Include in configdocs only the schemas for the current version.

Don't include schemas from other versions in the jars and distribution.

Include the spring .xsd in the distribution (it will be uploaded to the website from the distribution dir as well)

Remove old schema files and references to them *-7.0.xsd

Include the spring schema in configdocs/