ISPN-6576 Don't skip cache loader load for functional write-only and read-write commands on non-primary owners.
The functional write commands are not like all the other write commands: whereas
regular write commands are executed on one owner and the change is replicated to
other owners, the functional commands are supposed to be executed on all the
owners so that further replication is not necessary. For that to work we need to
allow the cache loader to load the value for these commands also for non-primary
owners.