ISPN-6576 Add FunctionalDistributionTest.
The scenario is roughly explained in the comment in the test's constructor: if
we execute the command on the secondary owner, the command should be skipped
there and forwarded to the primary owner who forwards it again to all secondary
owners, including the originator. If we get things wrong, like wrapping the
entry unnecessarily or loading (and wrapping) from cache loader unnecessarily,
the command will be executed when (and where) it should not be. This concerns
the Functional API more because the functions are not expected to be idempotent
like the other write commands are.

This test will guard against a mistake I made earlier in the development of this
patch series, whereas I misunderstood how the commands are forwarded back and
forth between originator, the primary owner and the secondary owners, causing
the command to be executed twice on some nodes in some scenarios.