ISPN-6706: Do not invoke CacheWriter.purge() if there is no eviction and the writer has preload enabled.
The thinking behind this is that when these two conditions are met, the cache
store's state is expected to be an exact copy of the transient state, and
purging the store itself is redundant: since the transient state is expected to
be the same as the persistent state, the entry will be removed from the store
IFF it is removed from the transient state.

Caveat: manual calls to evict() will break this assumption.