ISPN-6755: make MetaParams publically-instantiable.
These are the meta parameters used by the Functional API. They are created and
updated when we call EntryView.WriteEntryView.set(value, params...), but this is
not the only source of these parameters. The other is the cache loaders: the
return value of CacheLoader.load() is a MarshalledEntry that contains three
values: the key, the value and the metadata. This is a place where the
implementations will want to inject information like the lifetime (for
expiration). The only way to do that currently is to create a custom
implementation of the Metadata interface, but that won't work, partialy because
the Functional API is unfinished in these darker, dustier corners. So when we
call MetaParam.Lookup.findMetaParam(Class<T>), the only supported Metadata
implementation is the MetaParamsInternalMetadata (which external classes can't
create because all the constructors of MetaParams are package-private), and
there is a TODO to add interoperability support.

I think an easier way is to just allow external classes to instantiate
MetaParams so that we can create MetaParamsInternalMetadata which we can return
from CacheLoader.load()