ISPN-6910 Random failures due to MagicKey
Cache hashcodes of MagicKeys, and make sure they're unique.

Motivation: there was a fixed 1000-iteration loop to find out proper key. In tests that create many magic keys, the cumulative probability of failing one of the keys' generation was going up and this was causing failures in the testsuite. Therefore this was replaced by an algorithm that finds out a segment for given owners, and then finds a hashcode that maps to this segment. The hashcode is then cached and used for all keys in given segment.

As keys started to collide more often, an unique (counter-generated) number was added to each magic key for equality checks.