ISPN-5424 Replace SemaphoreCompletionService with LimitedExecutor
Instead of having an explicit continueInBackground method, the new
class allows a task to return a CompletableFuture, keeping the permit
unavailable until the CompletableFuture is done.

Introduce a "blocking mode" for both LimitedExecutor and
SemaphoreCompletionService: when they are used with a
WithinThreadExecutor, the task is not added to the queue, and instead it
is executed on the current thread. This is needed for total order
caches, so that topology updates block the total order thread.