ISPN-7116 Test methods that use a TestNg dataProvider are duplicated in JUnit reports
https://issues.jboss.org/browse/ISPN-7116

I was unable to find the class that writes the current
"TEST-<suite>.xml" file, so I created a new reporter. I know it isn't
ideal to have two files generated, but I need this fix to proceed with
Polarion onboarding. The JIRA can be left open, if this solution isn't
satisfactory.
Also updated TestNGTestListener to output the parameters to the console