ISPN-6997 Do not change protocols when the cluster changes size
UNICAST3 and NAKACK2 each guarantee ordering, but only for messages
sent with the same protocol.
In particular, each protocol can retransmit messages after a merge,
but the number of retransmitted messages depends on whether the
connection was closed in UNICAST3, and on the last STABLE run for
NAKACK2. So it is possible for a node to receive a LockControlCommand
sent with UNICAST3, but not a CommitCommand sent later with NAKACK2.