ISPN-8186 Avoid client socket connection to itself
* When a server stops, the client can still try to establish a
connection with a closing server until the new view has been
received.
* When opening the connection, there's a chance that
the client port and the local port are the same (since the
server por might have been vacated while stopping.
* The fix proposed verifies that if a connection is established,
the port and local port are not the same, in which case the socket
is closed and retried.