ISPN-8538 TwoWaySplitAndMerge[DIST_SYNC] deadlock in conflict resolution
Deadlock caused by StateReceiverImpl::stop synchronizing on
StateReceiverImpl.this, but then attempting to cancel a
StateRequestImpl.SegmentRequest whoms cancel() method synchs on itself.
However, a SegmentRequest can also be cancelled by another thread
when a CompletableFuture completes exceptionally. This results in
Thread-A acquiring StateReceiverImpl.this in StateReceiverImpl::stop and
calling SegmentRequest::cancel, however Thread-B has called SegmentRequest::cancel and
is now waiting to synchronize on StateReceiverImpl.this, hence deadlock.

Also, we shouldn't call StateReceiverImpl::stop in the
DefaultConflictManager.ReplicaSpliterator::tryAdvance when a
Cancellation/Execution Exception occurs as failed requests are already
cancelled on an exception in StateReceiverImpl.