Added mandatory parameter in configuration example
Hi,

If the `password` attribute is not given, connection pool provider will try to connect to the db with a null password, which will throw an exception if a password exists for that db. I think it would be good idea to add this parameter in the example ,just as a precaution so that no one gets doubt.