ISPN-10591 Replace UserMarshallerBytes with MarshallableUserObject wrapper
Allows user objects to be marshalled with the configured user marshaller
by a custom Protostream Marshaller implementation when calls to
ProtobufUtil are made. This means that it's no longer necessarry to
generate a byte[] using the configured user marshaller and writing this
as a byte[] protofield, which means exposing the configured user
marshaller to many Pojos. Instead, the calls to the configured user
marshaller are limited to the MarshallableUserObject marshaller.