ISPN-10591 Require JBossUserMarshaller to be explicitly configured
Automatically loading JBossUserMarshaller when
`infinispan-jboss-marshalling` is on the classpath is too restrictive,
because it means we can't cater for situations where we want
jboss-marshalling to be available, say for transcoding legacy clients, but we still
want a protostream user marshaller.

GlobalMarshaller no longer loads ExternalJbossMarshaller and
JBossMarshaller if on cp. This only changes the behaviour of nested
object lookups when utilisng Serialization, however loading these
implementations would no longer be possible when the GlobalMarshaller
becomes protostream based anyway.