ISPN-13264 Allow blocking in ChannelFactory methods
The locks are normally acquired for a short time as
even the opening and closing are non-blocking.
But it's still possible to receive concurrent topology updates
and have one update park to wait for another to finish.
Downgrading to a read lock while opening or closing channels
should reduce the possibility, but won't eliminate it.