ISPN-12712 Re-enable DuplicatedDomainsCdiIT
Remove the cache get operation, since the test is only supposed
to test whether the CDI extension sets the cache manager name or not.

In the WildFly test, import meta-inf from the Infinispan module.
Weld requires meta-inf in the manifest entry to see the Infinispan
beans that are not created explicitly by InfinispanExtensionEmbedded.