ISPN-13430 IRAC tombstones are leaking
Tombstones are required to detect conflicts with removed keys and they
can be removed when there is no pending update from any site.

In this PR, the primary owner queries all sites and, if all sites do not
have any update pending for that key, it removes the tombstone from the
primary owner and backup owners.