ISPN-10706 StateTransfer should complete segments as they are transferred.
Changing OutboundTask to be a Flowable

Changing the outbound task to be a flowable instead of a completable,
also changing the `StateProviderImpl` and the scattered one to handle
the flowable. Using the `LocalPublisherManager` to get the entries for
the segments.

Added the flag `STATE_TRANSFER_PROGRESS` to identify when the state
transfer is progress. This is applied to the command to retrieve the
cache entries during the state transfer.

Avoiding the remote fetch for remote/anchored entries during state
transfer. If the remote entry is requested the `AnchoredStateTransfer`
could change the remote address owner, and there is no need to remotely
retrieve the entry for state transfer.