ISPN-9366 Cache.size optimizations
* Counting the nuber of entries that have a TTL in the data container.
We use to verify if there are entries that can expire. We update the
value when the underlying map updates.

* Created a method for the size in the Publisher* interface and classes.

* Created in `PersistenceManager` a method for size while filtering the
segments to look.

* We query locally for the size and issue request to some of the remote
targets to do the same. With the returned value we sum everything.

* Change ComponentRegistry to create the LocalClusterPublisherManager,
only needed so the component is already started when used.

* Move the distributed execution up to ClusteringInterceptor while
executing only locally in the CallInterceptor and in
CacheLoaderInterceptor.