ISPN-13828 EventLogger for finished rebalance
Using the `EventLogger` in the `LocalPublisherManagerImpl` to log about
state transfer, when it starts and when it ends.

Added a view listener to the local topology manager to log CLUSTER event
about joining and leaving nodes. Since that was present in the cluster
topology manager, I moved everything to a single class that both of them
uses.