ISPN-12350 Persistent UUIDs are only used for initial consistent hash
We need to use the `PersistentUUID` while hashing the node's address.
Initially, I was going to only pass the method from
`PersistentUUIDManager` in all methods in `ConsistentHashFactory`,
but soon enough I realized that was not a good approach.

To easily propagate the `PersistentUUID` between operations I created
the `Member` class to be used in the factories and the
`SyncConsistentHash` that hold the members but delegates most of the
functionalities to the `DefaultConsistentHash`.

In the factory I kept the existing methods and the new ones will
delegate to the existing. I've updated in the `ClusterCacheStatus` to
use the new methods.

* When the `SyncConsistentHash` is serialized the UUID are also included,
so the serialized hash can be larger when compared with the previous
version. From the `SyncConsistentHashFactoryTest` we also can see that
is possible that more segments are moved.