ISPN-14125 Preloader can execute multiple times
Enabling `PreloadManager` to run more than once.
We add a versioning to the manager, so it can keep track of loaded
versions. When the entry is loaded, it is associated with a numeric
version equal to the manager's. We also mark it as an entry loaded.
After loading all the entries we remove the ones with an older version.