ISPN-14329 Availability of caches should be prevented until a cluster is complete after "shutdown cluster"
After the cluster shutdown, the nodes must wait for the complete
recovery before accepting commands. We keep the cache's registry as
`INITIALIZING` until all nodes join back and the restoration of a stable
topology.

We log and return an exception with the current members. We can't say
the missing member's address for sure, as we only have the persistent
UUID. The coordinator throws the exception with the current member list.

Created a dedicated exception to handle missing members. Avoid logging the trace.

The downside is that SIFS and single-file store now creates the store
from the previous file, even when purge on startup is enabled, to
possibly clear all the data after. This could be improved further in
following PRs.