ISPN-14500 REST API blocking when retrieving keys and entries
* Instead of using the blocking `InputStream` we implement a
`ChunkedInput`. We have control over the stream has ended or not.

* Instead of using the `CacheStream` we use `ClusterPublisherManager`
and subscribe to the `Publisher` using the `DefaultSuscriber`. With
this subscriber, we have control on how much elements to consume.

* We try to fill the 8192 bytes for the data chunk, consuming as much elements
necessary. We might require more elements from the underlying publisher,
but we use a bounded queue, and never overflow its size.