ISPN-14477 Tests for concurrent Spring session access
Added more tests to cover the change tracking in `InfinispanSession`. Specifically, ensure that the session attributes, max inactive duration and last accessed time are updated in the Infinispan cache if they are changed. Also, added tests for the various FlushMode and SaveMode settings.

Fixed a `NullPointerException` in `SessionUpdateRemappingFunctionProtoAdapter` which occurred when a session attribute was removed, i.e. the `SessionAttribute` value was null.

Fixed a serialisation issue `SessionUpdateRemappingFunctionProtoAdapter` which caused the last accessed time to be set to the epoch value if the session's last accessed time hadn't been changed.