ISPN-15357 Handling stateless nodes during cache join
This PR has two sides. First, by "stateful" we refer to nodes with
information in the persistent state about the cache cluster before
restart, and "stateless" a node without any information.

Originally, a stateful cluster only accepts stateful joiners until the
old cluster is complete. Stateless nodes can only join *after* that, but
they have no clue about that, on what stage the recovery is. If they try
joining too early, the receive an exception, which could lead to the
node failing to start.

With the changes, a stateless node can issue a join request at any
point. It will *not* join the cluster imediatelly, as the cluster needs
to recover, but now the node receive a specific response, so it can send
a join request again after the recovery is complete. This is automatic
now, and shouldn't need any intervention.

The tricky case, starting a stateless coordinator with a stateful
joiner. This is very likely a problem during initialization, or a
misconfiguration during updates. In such a case, the stateful node
trying to join will receive an exception, and likely fail to start. This
would possible identify issues elsewhere and the user would need to
manually fix it.

If the user is manually adding nodes, it needs to make sure to start the
new node *after* all the previous nodes are up and running. Any
concurrent cluster start, or concurrent membership changes (i.e., a node
shutdown while adding another one) could trigger this behavior.