ISPN-15494 Near cache with invalidation fails to start in Spring Boot
Updated for the invalidated near cache instances. We check if the
MANAGER marshaller is capable of handling the server type, and only then
proceed with the usual flow. Either we register listener and set data
format, or the inverse.

If the marshaller is compatible with the server type, we exchange the
type information when installing the listener. See, this might hide bugs
that bite back during runtime! This might be found during runtime when
executing some key based operation against the server.

For internal caches, where the user won't interact directly through HR
(resp), this is fine. We also have a warning in any case.