ISPN-15586 Concurrent starting nodes can hang view installation
Added a test covering a few scenarios. The fix involves not waiting the
LocalTopologyManager to start. Once the component enter into running
mode, it will request the topology information from the coordinator. We
can avoid these initial messages before the initialization.

The problem does not happen when the global rebalance is disabled. I
believe the coordinator does not multicast the topology change in this
case and there is no message blocking.