[#14313] Metrics with manually created JGroups channel
Currently if a JGroups channel is created manually then no metrics will
be registered for it since the metrics channel listener is not added.
This change makes it so that the channel listener is added when the
channel is initialized regardless of how it is created (it also
contains some minor cleanup and fix for warnings).
Closes #14313