[#14315] Allow disabling base metrics
When running in embedded mode, it is desirable to disable the JVM
metrics from being exported by Infinispan.
Such metrics are (or should be) exported by the application framework
(e.g. Spring), not by the different libraries (infinispan) included by
an application.
This change adds a setting to disable JVM metrics, and deprecates the
existing base binders in favor of Micrometer binders, with a setting to
disable the legacy metrics.
Closes #14315