[#16001] Show unknown auth mechanism name instead of 'null'
If using the REST API authorization mechanism that isn't enabled in the
Infinispan server, the error message was an unhelpful: "ISPN080052: The
request authentication mechanism 'null' is not supported"

This is because it was using the mechanism name retrieved from the
challengePrefixMechanisms map, and if a mechanism isn't enabled, that
map will return null. This fixes the problem by printing the mechanism
name sent by the client instead.